requirements.txt
backoff==1.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
charset-normalizer==2.0.10; python_full_version >= "3.6.0" and python_version >= "3"
deprecated==1.2.13; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
googleapis-common-protos==1.54.0; python_version >= "3.6"
grpcio==1.43.0; python_version >= "3.6"
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
opentelemetry-api==1.8.0; python_version >= "3.6"
opentelemetry-exporter-otlp-proto-grpc==1.8.0; python_version >= "3.6"
opentelemetry-instrumentation-requests==0.27b0; python_version >= "3.6"
opentelemetry-instrumentation-urllib3==0.27b0; python_version >= "3.6"
opentelemetry-instrumentation==0.27b0; python_version >= "3.6"
opentelemetry-proto==1.8.0; python_version >= "3.6"
opentelemetry-sdk==1.8.0; python_version >= "3.6"
opentelemetry-semantic-conventions==0.27b0; python_version >= "3.6"
opentelemetry-util-http==0.27b0; python_version >= "3.6"
protobuf==3.19.3; python_version >= "3.6"
psycopg2-binary==2.9.3; python_version >= "3.6"
python-dotenv==0.19.2; python_version >= "3.5"
requests==2.27.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
urllib3==1.26.8; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0" and python_version < "4")
wrapt==1.13.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
